more sensical ipv{4,6} implementations
- reworked constructors
- better unit tests
- windows still sucks
- yatta yatta yatta
