+++ tests/utils.cpp
@@ -165,46 +165,4 @@ namespace oxen::quic
         logger_config(out, type, lvl);
     }
 
-    std::pair<std::string, uint16_t> parse_addr(std::string_view addr, std::optional<uint16_t> default_port)
-    {
-        std::pair<std::string, uint16_t> result;
-        if (auto p = addr.find_last_not_of("0123456789");
-            p != std::string_view::npos && p + 2 <= addr.size() && addr[p] == ':')
-        {
-            if (!parse_int(addr.substr(p + 1), result.second))
-                throw std::invalid_argument{"Invalid address: could not parse port"};
-            addr.remove_suffix(addr.size() - p);
-        }
-        else if (default_port)
-        {
-            result.second = *default_port;
-        }
-        else
-        {
-            throw std::invalid_argument{"Invalid address: no port was specified and there is no default"};
-        }
-
-        bool had_sq_brackets = false;
-        if (!addr.empty() && addr.front() == '[' && addr.back() == ']')
-        {
-            addr.remove_prefix(1);
-            addr.remove_suffix(1);
-            had_sq_brackets = true;
-        }
-
-        if (auto p = addr.find_first_not_of("0123456789."); p != std::string_view::npos)
-        {
-            if (auto q = addr.find_first_not_of("0123456789abcdef:."); q != std::string_view::npos)
-                throw std::invalid_argument{"Invalid address: does not look like IPv4 or IPv6!"};
-            else if (!had_sq_brackets)
-                throw std::invalid_argument{"Invalid address: IPv6 addresses require [...] square brackets"};
-        }
-
-        if (addr.empty())
-            addr = "::";
-
-        result.first = addr;
-        return result;
-    }
-
 }  // namespace oxen::quic
